+++ CHANGELOG.md
@@ -12,6 +12,7 @@ NOTES
 FEATURES
 
 * **New Check:** `AT005`: check for acceptance test names missing `TestAcc` prefix
+* **New Check:** `AT006`: check for acceptance tests containing multiple `resource.Test()` invocations
 * **New Check:** `S020`: check for `Schema` of only `Computed` enabled with `ForceNew` enabled
 
 ENHANCEMENTS